+++ adapter/make_mocha_tests_it.js
@@ -12,13 +12,22 @@ module.exports = function makeMochaTest(tests) {
         }
     }
 
+    function getResult(currentTest) {
+        return {
+            duration: currentTest.duration,
+            message: currentTest.err?.message,
+            status: currentTest.state,
+            exception: currentTest.err
+        }
+    }
+
     function executeStepByText(text, argument) {
         const steps = supportCodeLibrary.stepDefinitions
             .filter(stepDefinition => stepDefinition.matchesStepName(text));
         if (steps.length === 0) throw new Error(`Step '${text}' is not defined`);
         if (steps.length > 1) throw new Error(`Step '${text}' matches multiple step definitions`);
-        const [ step ] = steps;
-        const { parameters } = step.getInvocationParameters({
+        const [step] = steps;
+        const {parameters} = step.getInvocationParameters({
             step: {text, argument},
             world: this
         });
@@ -75,15 +84,16 @@ module.exports = function makeMochaTest(tests) {
         }
     });
 
-    afterEach( function () {
+    afterEach(function () {
         const test = findTest(tests, this.currentTest.title);
         const world = this.world;
+        const result = getResult(this.currentTest);
         for (const afterTest of supportCodeLibrary.afterTestCaseHookDefinitions) {
             if (afterTest.appliesToTestCase(test)) {
                 runStep(afterTest.name, function () {
                     afterTest.code.apply(world, [{
                         pickle: test,
-                        result: null,
+                        result,
                         gherkinDocument: tests,
                         willBeRetried: false
                     }]);
@@ -99,7 +109,7 @@ module.exports = function makeMochaTest(tests) {
                 const stepName = keyword(step) + ': ' + step.text;
                 runStep(stepName, function () {
                     this.step = step;
-                    this.currentTest = { state: 'passed' };
+                    const result = { status: 'passed', duration: 0 };
                     for (const beforeStep of supportCodeLibrary.beforeTestStepHookDefinitions) {
                         if (beforeStep.appliesToTestCase(step)) {
                             beforeStep.code.apply(world, [{
@@ -116,7 +126,7 @@ module.exports = function makeMochaTest(tests) {
                                 pickle: test,
                                 pickleStep: this.step,
                                 gherkinDocument: tests,
-                                result: this.currentTest.state
+                                result
                             }]);
                         }
                     }