+++ CHANGELOG.md
@@ -14,6 +14,9 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 :microscope: - experimental
 
+## [1.4.1]
+- :rocket: fixed `result` property in `After` and `AfterStep` hooks
+
 ## [1.4.0]
 - :rocket: added `Template` utility function
 ```typescript